:hammer: Tratamento de erros (findOne, update e remove)

+++ src/products/products.controller.ts
@@ -6,6 +6,8 @@ import {
   Patch,
   Param,
   Delete,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -26,17 +28,41 @@ export class ProductsController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.productsService.findOne(+id);
+  async findOne(@Param('id') id: string) {
+    const product = await this.productsService.findOne(+id);
+    if (product === null) {
+      throw new HttpException(`ID ${id} was not found.`, HttpStatus.NOT_FOUND);
+    }
+    return product;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
-    return this.productsService.update(+id, updateProductDto);
+  async update(
+    @Param('id') id: string,
+    @Body() updateProductDto: UpdateProductDto,
+  ) {
+    const updateProduct = await this.productsService.update(
+      +id,
+      updateProductDto,
+    );
+    if (!updateProduct) {
+      throw new HttpException(
+        `ID ${id} was not found for change.`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    return updateProduct;
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const removeProducts = await this.productsService.remove(+id);
+    if (!removeProducts) {
+      throw new HttpException(
+        `ID ${id} was not found to be removed.`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
     return this.productsService.remove(+id);
   }
 }
